requirements.txt
Django==5.1.4
PyJWT==2.10.1
